xStick DTIM support

+++ qwifi/src/wlan_drv/qapi_wlan_param_group.c
@@ -388,8 +388,8 @@ qapi_Status_t qapi_WLAN_Get_Param(uint8_t __attribute__((__unused__)) device_ID,
             break; /* __QAPI_WLAN_PARAM_GROUP_WIRELESS_RSSI */
         }
         case __QAPI_WLAN_PARAM_GROUP_WIRELESS_STA_LISTEN_INTERVAL_IN_TU: {
-            uint32_t *interval = (uint32_t *)data;
-            if (*length < sizeof(uint32_t)) {
+            uint16_t *interval = (uint16_t *)data;
+            if (*length < sizeof(uint16_t)) {
                 return QAPI_WLAN_ERR_EINVAL;
             }
             wlan_get_sta_slptime(interval);
